Fix how prop types for SearchResults are defined
