+++ webapp/channels/src/components/search_results/search_results.test.tsx
@@ -7,13 +7,12 @@ import React from 'react';
 import {DATE_LINE} from 'mattermost-redux/utils/post_list';
 
 import SearchResults, {arePropsEqual} from 'components/search_results/search_results';
+import type {Props} from 'components/search_results/search_results';
 
 import {renderWithContext} from 'tests/react_testing_utils';
 import {getHistory} from 'utils/browser_history';
 import {popoutRhsSearch} from 'utils/popouts/popout_windows';
 import {TestHelper} from 'utils/test_helper';
-
-import type {Props} from './types';
 
 jest.mock('utils/popouts/popout_windows', () => ({
     popoutRhsSearch: jest.fn(),
@@ -117,6 +116,7 @@ describe('components/SearchResults', () => {
         setSearchFilterType: jest.fn(),
         updateSearchTeam: jest.fn(),
         updateSearchTerms: jest.fn(),
+        dispatch: jest.fn(),
     };
 
     beforeEach(() => {
