+++ webapp/channels/src/components/search_results/search_results.tsx
@@ -24,6 +24,7 @@ import NoResultsIndicator from 'components/no_results_indicator/no_results_indic
 import {NoResultsVariant} from 'components/no_results_indicator/types';
 import DateSeparator from 'components/post_view/date_separator';
 import {getSearchPopoutTitle} from 'components/rhs_search_popout/title';
+import type {SearchFilterType} from 'components/search/types';
 import SearchHint from 'components/search_hint/search_hint';
 import SearchResultsHeader from 'components/search_results_header';
 import LoadingWrapper from 'components/widgets/loading/loading_wrapper';
@@ -38,11 +39,32 @@ import FilesFilterMenu from './files_filter_menu';
 import MessageOrFileSelector from './messages_or_files_selector';
 import PostSearchResultsItem from './post_search_results_item';
 import SearchLimitsBanner from './search_limits_banner';
-import type {Props} from './types';
+
+import type {OwnProps, PropsFromRedux} from './index';
 
 import './search_results.scss';
 
 const GET_MORE_BUFFER = 30;
+
+export type Props = OwnProps & PropsFromRedux & {
+    channelDisplayName?: string;
+    crossTeamSearchEnabled: boolean;
+    isCard?: boolean;
+    isChannelFiles: boolean;
+    isFlaggedPosts: boolean;
+    isMentionSearch: boolean;
+    isOpened?: boolean;
+    isSideBarExpanded: boolean;
+    searchFilterType: SearchFilterType;
+    searchType: SearchType;
+
+    getMoreFilesForSearch: () => void;
+    getMorePostsForSearch: () => void;
+    handleSearchHintSelection?: () => void;
+    setSearchFilterType: (filterType: SearchFilterType) => void;
+    updateSearchTeam: (teamId: string) => void;
+    updateSearchTerms: (terms: string) => void;
+};
 
 interface NoResultsProps {
     variant: NoResultsVariant;
